requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
annotated-types==0.5.0
anyio==3.7.1
async-timeout==4.0.3
attrs==23.1.0
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.6
exceptiongroup==1.1.3
fastapi==0.103.1
food2vec==0.0.4
frozenlist==1.4.0
gensim==4.3.2
h11==0.14.0
idna==3.4
joblib==1.3.2
multidict==6.0.4
openai==0.28.0
pandas==2.1.0
pybind11==2.11.1
pydantic==2.3.0
pydantic_core==2.6.3
python-dateutil==2.8.2
pytz==2023.3.post1
requests==2.31.0
scikit-learn==1.3.0
scipy==1.11.2
six==1.16.0
smart-open==6.4.0
sniffio==1.3.0
starlette==0.27.0
threadpoolctl==3.2.0
tqdm==4.66.1
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.4
uvicorn==0.23.2
yarl==1.9.2
torch==2.0.0
torchvision==0.15.1
numpy==1.25.2
matplotlib==3.4.3
Pillow==8.4.0
python-multipart==0.0.5